test: split App.test.tsx into 4 files under 200 LOC each
Splits the 549-line App.test.tsx into parseLogLine.test.tsx (154),
extractionHelpers.test.tsx (107), deriveProviderHealth.test.tsx (149),
and formatHelpers.test.tsx (131) per Constitution Rule 7.

Aloop-Agent: build
Aloop-Iteration: 2
Aloop-Session: orchestrator-20260321-172932-issue-189-20260323-151506
